requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
python-slugify==8.0.4
requests==2.32.1
sqlparse==0.5.0
stripe==9.7.0
text-unidecode==1.3
typing_extensions==4.11.0
urllib3==2.2.1
